cleanup AudioInjector memory management
